+++ r/R/RcppExports.R
@@ -725,6 +725,10 @@ checksum_double_cpp <- function(matrix) {
     .Call(`_BPCells_checksum_double_cpp`, matrix)
 }
 
+apply_matrix_double_cpp <- function(mat_sexp, f, row_major) {
+    .Call(`_BPCells_apply_matrix_double_cpp`, mat_sexp, f, row_major)
+}
+
 matrix_identical_uint32_t_cpp <- function(mat1, mat2) {
     .Call(`_BPCells_matrix_identical_uint32_t_cpp`, mat1, mat2)
 }